+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 All notable changes to EMU will be documented in this file.
 
+## [4.0.2] - 2026-05-24
+### Added
+- Added `emass upload artifact`
+
+### Changed
+- Converted the `contentType` parameter into a map-based `headers` parameter in the `emass.POST` function
+
+### Removed
+- Comments that weren't adding value
+
 ## [4.0.1] - 2026-05-24
 ### Added
 - `emu test api` subcommand for testing connectivity to the eMASS API
@@ -14,5 +24,6 @@ All notable changes to EMU will be documented in this file.
 ### Changed
 - Migrated from v3 configuration format
 
+[4.0.2]: https://github.com/deathlabs/emu/releases/tag/v4.0.2
 [4.0.1]: https://github.com/deathlabs/emu/releases/tag/v4.0.1
 [4.0.0]: https://github.com/deathlabs/emu/releases/tag/v4.0.0